requirements.txt
httpx==0.28.1
PyNaCl==1.5.0
python-dotenv==1.1.1
python-telegram-bot==22.1
